fix: parseFloat instead of parseInt for threshold

+++ src/inputHelper.ts
@@ -61,7 +61,7 @@ export async function getInputs(): Promise<IInputs> {
       .getInput('sfl-threshold', {required: true})
       .replace(/\[|\]/g, '')
       .split(',')
-      .map(value => parseInt(value))
+      .map(value => parseFloat(value))
   } catch (error) {
     throw new Error(
       'Invalid form of input `sfl-threshold`. It should be a comma separated list of numbers.'
